i156 Revert back to using the fully-qualified name as batch artifact name for org.hibernate.search.jsr352.JobContextSetupListener
We can't use the batch.xml to provide our own artifacts, because in
standard, JavaSE applications, the user may provide its own batch.xml,
which will erase our own and make the reference to the job context setup
listener undefined.

We work around this situation by using the fully-qualified class name as
a batch artifact name, and by naming our CDI bean with this exact name.
It's a bit strange to have a bean name be a fully-qualified class name,
but it works.